requirements.txt
fal-client
requests
Pillow
numpy
torch
